requirements.txt
fastapi
gradio
uvicorn
transformers
gtts
deep-translator
pillow
speechrecognition
